bug fix in case of nil persistence manager is provided

+++ state_machine.go
@@ -111,12 +111,15 @@ func (machine *StateMachine) Run(isDelete bool) error {
 			return err
 		}
 
-		// ask persistence manager to save the state
-		err = machine.persistenceManager.Save(data)
+		// check if persistence manager is given
+		if machine.persistenceManager != nil {
+			// ask persistence manager to save the state
+			err = machine.persistenceManager.Save(data)
 
-		// check if persistence manager has reported any errors
-		if err != nil {
-			return err
+			// check if persistence manager has reported any errors
+			if err != nil {
+				return err
+			}
 		}
 
 		// execute machines current state
@@ -124,7 +127,7 @@ func (machine *StateMachine) Run(isDelete bool) error {
 	}
 
 	// all execution has finished, check if state delete is required
-	if isDelete {
+	if isDelete && machine.persistenceManager != nil {
 		// delete the machines state as its not required any more
 		machine.persistenceManager.Delete()
 	}
